moving EtradeAuthResponse into its own file
